Home and About components working with Angular routing
Changed from electron to express. A few lessons learned here. 1) In the
main.js, you have to use "*" to return all paths to the single page
application to spoof the routes. 2) Need to use the factory for methods
because executing on page load doesn't work with angular. 3) Electron
sucks for web-page design. 4) Cannot navigate to other modules. Instead,
they should be added as direct templates instead of using a templateUrl
for routing.